add delete confirmation

+++ public/build/js/core.js
@@ -399,11 +399,13 @@ function convertClaimTax(value) {
 }
 
 function remove(address) {
-    storeAccounts.splice(storeAccounts.indexOf(address), 1)
-    delete storeNames[address]
-    if (storeAccounts) localStorage.setItem('accounts', JSON.stringify(storeAccounts))
-    if (storeNames) localStorage.setItem('names', JSON.stringify(storeNames))
-    refresh()
+    if (confirm(`Are you sure you want to remove ${storeNames[address]}?`)){
+        storeAccounts.splice(storeAccounts.indexOf(address), 1)
+        delete storeNames[address]
+        if (storeAccounts) localStorage.setItem('accounts', JSON.stringify(storeAccounts))
+        if (storeNames) localStorage.setItem('names', JSON.stringify(storeNames))
+        refresh()
+    }
 }
 
 async function simulate(address) {
